Updated default value setting
Let's say I want to adjust the default duration cap from 60 to 45, before it would crash if the value was say 50, because it is now outside of the range. Now it checks if the value loaded by appstorage is within the appropriate range.  maybe this logic should be moved to the overall obs. obj. instead of the building block views...